Translated 'While_Valid_1' test case. This upgrade has two change: void type and overlapping.
The void type is used to declare an empty list in Whiley, e.g. 'newlist
%6 = () : [void]'. As the void type has no mapping type in C, we convert
the void type into integer type. Alos, no translation is output for this
new list byte-code.

Overlapping is strictly forbidden in C so the function of printing an
array is renamed as 'indirect_printf_array'