Remove the redundant BoundChange class and hashmap.
In order to keep track of bound changes for each variable, Domain class
is added with 'isUpperBoundIncreasing' and 'isLowerBoundDecreasing'
boolean flags. And Bounds class is implemented with'checkBoundChange'
and 'WidenBounds' to check bound change for each iteration and widen
upper/lower bounds if the upper/lower bound of a variable increases or
decreases with iterations.