Added the function to compute in set for each block.
The 'computeIn' takes an 'in' set and a list of byte-code, to compute
the new in set by using the live analysis of Whiley compiler, and then
returns the computed 'in' set as the output. 

This function allows to recursively compute live variables inside an
assertion or an invariant and overwrite 'in' set.