Generated C code with liveness analysis results.
If 'copy' option is on, then the code generator runs the copy
elimination analysis to produce the liveness, and while translating the
invoking byte-code, the generator asks the analyzer whether the array is
alive in the following block. If the array is live, then it requires a
clone. Otherwise, the array can be safely overwritten with the return.