Solved the inefficiency problem caused by processbuilder.
I noticed that using processbuilder to run analysis is very slow due to
a large amount of output messages. This is needed for debugging but
causes a problem on testing. So the analyzer outputs the analysis
results only when the 'verbose' option is enabled, so that the messages
can be reduced to a minimum and the performance of our project can be
increased.