Added type casting while generating the code of array comparison.
While translating the code of comparing two arrays, the code generator
checks the LHS or RHS is an array of integers as some arrays are of void
type. For example, the Wyil code
 
	ifeq %1, %8 goto blklab6 : [int]

can be translated into C code:

	if(isArrayEqual(_xs, _xs_size, (long long*)_8, _8_size)==1){
		goto blklab6;
	}

Register 8 is casted from an array of voids into an array of integers.