Verify the exit value of generated C code.
This test case suite uses the code generator to translate a whiley
program into C code (with/without copy optimization). Then it compiles
and runs the generated code, and validate the exit value of the program.
If the value is not 0, throw out assertion failures. This change ensures
that the generated C code can be compiled and executed correctly.