Updated built-in 'convertArgsToIntArray' and remove sysout files.
This update aims to avoid the dynamic allocation for the return array.
The program computes the array size first and then allocates the array
with given size and gradually fills in each element. Also, remove all
the system out files inside 'tests\code' folder due to the new update of
code generation validation.