Validating the generated code of valid test cases.
New test cases each reads a valid Whiley program, and generates and
places the C program inside a subfolder of 'tests/code' folder with test
case name as folder name. The naive C program is put inside 'slow'
folder and copy-optimized C program is placed in 'fast' folder. 

The benefits of this change

1. No needs of copying valid Whiley programs.
2. Generated C code is placed with respect to the optimization.