Added 'sort_call_by_reference' test code for code generator.
This test case tests the call_by_reference merge sort method. This new
method passes the input array as argument pointer, rathe than an array,
and manipulate the input array directly. As being passed as pointer, the
changes to array can be visible outside the method. The
call-by-reference method avoids the array copies of function calls and
reduce the memory usage.