Changed C compilation command (from 'gcc' to 'gcc-4') on Windows.
This update changes the C compilation commands, sent from Java process,
from 'gcc' to 'gcc-4'. On windows, the command prompt can not recognize
'gcc' command (it is a symbol link in Cygwin), so we need to use actual
file name (gcc-3 or gcc-4) to call GCC and compile C programs. Note that
Cygwin path (e.g. 'c:\cygwin\bin') needs to be added to PATH variable.