Added 'size' variable for a user-defined structure.
In TicTacToe test case, 'Board' structure contains an integer array
field 'pieces'. In order to indicate the length of 'piece' field, we add
a size variable (pieces_size). This change also applies to newrecord
wyil code. When initializing a Board structure, we assign the array to
'pieces' field but also gives the array size to 'pieces_size' field.