Updated copy reduced C code for TicTacToe test case
To access an array member in a structure, we need to use a temporary
variable to point to the array member, e.g. 

int[] piece = b.pieces;

This assignment does not require the array copy as 'piece' variable is
acted as a temporary variable to access the member array 'piece' of b
structure.