freed 2D array and removed the array copy inside 'matrix' function.
To free a 2D array, we need to loop over top-leveled array and free each
sub-array, and otherwise we have the memory leaks for sub-array, e.g. 
	long long** data;
	
	for(i=0;i<size;i++){
		free(data[i]);
	}
	
	free(data);

The other change is to remove the un-necessary array copy in 'matrix'
function as the input 'data' 2D array is read-only and not changed by
the function.