Removed the copy of structure member in field load code
The copy analyzer checks if the member (field) of a given structure is
being mutated in the function to remove the copy for fieldload code. If
the member is read-only, the copy can be safely removed, e.g. 
	'fieldload %12 = %1 data : {int[][] data,int height,int width}'
can be translated into
	'_12 = A.data;'
where _12 is a read-only variable in the function.