Generated 'null' member of a structure
The extra 'null' members are added to a structure to indicate whether
the structure is empty or not, e.g. 'type Board is (null | {nat move,
Square[] pieces} board)' can be translated into C code:

typedef struct{
	void* null;
	Square* pieces;
	long long pieces_size;
	nat move;
} Board;
where the extra pointer 'null' stores the address of structure if the
board is not empty.