Assigned ownership value for rhs variables of an assignment
This update aims to give the ownership value for the rhs assignment
variable. If the copy of rhs variable is needed, then the ownership is
true. Otherwise, the ownership is false. For example, 'a = copy(b)'
assignment means 'a' and 'b' both have the ownership whilst the pure
assignment 'a = b' means 'a' have the ownership but 'b' does NOT.