Passed array parameters using '_DARRAY_PARAM' marcos
The input parameters can be directly passed or copied and then passed to
the function call. The '_XDARRAY_PARAM' marcos takes an array variable
and produces a parameter string that contains array and size variables.
And the '_XDARRAY_COPY_PARAM' makes a copy of the array variable and
produces the parameter string with size variable.