Fixed the issued related to a struct pointer return value.
The 'EmptyBoard' in newTicTacToe test case creates a local board and
returns the board in a form of structure pointer ('Board*'). This would
cause the memory access inconsistency issue as local boards are
allocated in internal memory but structure pointers are constructed in
raw memory spaces. 

The solution is that the local board is dereferenced and copied onto the
raw memory space using 'malloc' function, and then the pointer to the
new board is passed as the returned value of 'EmptyBoard' function.   



 pointer of new board is returned to the caller of 'EmptyBoard'
function.