Add 'computeOwnership' function to generated the ownership code
The new 'computeOwnership' function allows the deallocation analyzer to
take one wyil code and follows ownership rules to compute the
ownerships for each operand and then generate the corresponding code,
e.g. adding or removing ownerships. 

This change not only reduces the code generator's workload but makes a
good use of deallocation analyzer.