Computed the ownership value for the input parameter of a function call
The ownership value of input parameters are computed and passed into the
function call. To determine the parameter ownership, the de-allocation
analyzer checks if the parameter is mutated and returned by the called
function, and if the parameter is still alive after the function call.
By taking union of these results, the analyzer can infer the ownership
of input parameters.