Computed and passed ownership along with parameters to a function call
The ownership rules for a function call parameter is similar to that of
copy elimination. By checking mutation, return value and liveness, the
de-allocation analyzer can determine whether the parameter after the
function call still owns the memory or not. 

But this new change still needs benchmark tests to verify if the leaks
of generated C code are effectively stopped by new ownership rules.