Fixed bugs of translating union-typed parameter for a function call
The union-typed parameter, such as 'null|int', is used to pass the
command line argument into the program. As it represents an integer, the
parameter can be passed to the function call without any copies or
memeber access, e.g. 'max' is an union typed ('null|int') variable and
can be passed to the function call 'a = genMatrix(max, max)'.