Added the ownership rules for 'newlist' code.
The newlist code creates a new array with given elements, and thus its
lhs variable should be added to ownership set, e.g. 'newlist a = (b, c)'
declares 'a' as an array and initializes it with 'b' and 'c' values. In
this case, 'a' variable should be added to ownership set and released at
the end of function.