Added borrowing ownership
The callee can either borrow or own the parameter of a function call. If
the parameter is owned by the caller, then the callee borrows the
parameter, but does not need to free its memory space. Otherwise, the
parameter is transfered to the callee and its memory space is freed at
the end of function.