Added speedup benchmark results
The speedups show that copy and de-allocation analyzer can assist the
code generator to produce the most efficient C code and gain best
speedups in all test cases. For example, the speedups in MatrixMult test
case increase with matrix size and show that the efficiency of generated
C code can be improved by reducing the copying overhead and stopping
memory leaks. However, the speedups in TicTacToe and new TicTacToe test
cases decrease with problem sizes, and needs other approaches to
increase the efficiency.