Defined user types by using structure pointers in C
In Whiley, users can define types and set the variable to be NULL. But
when translating the user-defined types to C structure, we need to
define the structure and declare the variables to be structure pointers,
e.g. 'Matrix a;' in Whiley can be translated to
	'Matrix* a = malloc(sizeof(Matrix));'
where a is a structure pointer, pointing to a Matrix structure.