Added 'perfstat.sh' to profile and collect C program performance
'Perf' Linux command is used to measure the CPU usage of Matrix C
program with respect to 'GCC O3', 'Polly Optimization', 'multiThreading
OpenMP Code'. The preliminary results show polly compiler can
effectively parallelize none-function parameter handwritten C code, that
declares the matrix as global variables and does not pass it to the
function, but does not exploit the parallelization for pass-as-parameter
C MatrixMult program.