Adds base memory address in handwritten Matrixmult 2 c code
We adds a base address before the loop, i.e.
	 C[i*N+j] = 0; // Adds base address
     for(k=0;k<N;k++){
         C[i*N+j] = C[i*N+j] + A[i*N+k]*B[k*N+j];
     }
This change allows polly compiler to know and extract the base address
(C[i*N+j]) inside the loop, and optimize the loop. 

By changing this part of code, the execution time is decreased from 20
seconds to 2.1 seconds (7 times faster).