Modified 'VectorMult.c' to benchmark super large vector (137GB) mult
The 'VectorMult.c' aims to multiply two huge vectors, each containing
1024*1024*1024*16 elements and using more than 137 GB memory, and then
produces the dot product (17,179,869,184). 

Due to the memory capacity, the vectors can not be allocated in a single
machine. Thus, it splits the large-sized vector into a number of
fixed-sized (1024*1024) sub-vectors. Then it performs the vector
multiplication for each sub-vector and put the dot product to another
vector. And finally it sums up all the products in the vector to obtain
the final result.