Add a separate and nested loop to initialize random numbers
A nested loop is added to 'init' function to generate the random number for each Matrix element and stores the numbers with
a 'local' matrix. This change separate the random number generation from Matrix element assignment, so that Polly is able
to pipeline matrix assignment + matrix addition and gain the maximal parallelism.