Added 'autogenerate1' MatrixMult test case
The 'autogenerate1' test case takes command line arugments as matrix size to create input matrices and output resulting matrix. That means, the matrix size is determined at the runtime, rather than hard-coding sizes at the program, and varies with user input. Then we use Polly to optimize this program and, on our preliminary result, the generated C code still benefits the speedups from Polly.

The benchmarks are required to conform such a conclusion