Used the cached array to speed up the GCD Euclid's algorithm
The 'gcd_cached' function starts from 0 to N and takes two numbers to calculate the greatest commond divisor and stores all the found divisors with an array. This algorithm does not need the loop to decrease the remainders of two numbers, but performs the subtraction once and lower the numbers such that we can reuse the found gcds.