Added N-Queen test case
N-Queen problem place n queens on a n X n board so that no two queens
are on the same row, column or diagonal. This Whiley program simplifies
the original N-Queen benchmark program, and does not store all solutions
with an arrays but counts the total number of solutions instead.