Remove COPY at indexof code and remove ownership code to the assignment next to indexof code
This update in C code gets rid of the difference between GCC and Clang compiler. The indexof code access one element from an array and the assgnment next to it assigns the element to another variable. Within these two code, there would no copy or free or ownership transfer as shown
               a = b[i]; //The indexof code access 'b' array at 'i' index, and thus does not require copies
               c = a; // Assign a to c as a is the temporary variable, and thus does not need ownership