Transfer out the ownership of lhs variable that points to an array element
In byte-code level, accessing an array element requires the use of 'indexof' and 'assign' code, to take one element from array and assign it to a new variable. As the array is owned by the original array variable, accessing array does not need to make any copy nor transfer out any ownership. For example,

     a = b[i]; // Accessing the array at index i and loading it to variable 'a'
     a_has_ownership = false; // The temporary variable does not own the array
     c = a;   // Assign the array element to another variable
     c_has_ownership = false;
     a_has_ownership = false;
     // a or c does not have the ownership because the array is owned by b