Added special case to ownership rules for casting integer assignment
The assignment of casting an integer pointer to an integer should not make any change to the ownerships. For example, the assignment (int size = *max;) performs the dereference to obtain the value, and assigns the value to an integer variable. As lhs variable is an integer, lhs variable does not have the ownership. And rhs variable still keeps its ownership.