Upgrade the Whiley compiler to v0.3.40
This new change brings lots of changes on Whiley bytecode, and thus the
code generator and all analyzers requires to be rewritten to fit into
the new bytecode. At current stage, the project passes the compilation.
Also, the pattern matching is removed from our project due to the fact
that lists and while-loop is no longer used in Whiley.