Backed up LiveVariableAnalysis in wyil module
The livevariable analyzer of our project is referenced to the live
variables analyzer, developed by Dr. David Pearce, in wyil module. This
backup is because new wyil module infers the lifetime, rather than
analyzing liveness information, and thus does not provide the support
for this analyzer. To prepare the update, we back up the analyzer and
move it into our project code base.