Added 'postProcessor' to add post-deallocation code only
Previous to this update, the 'postProcessor' requires the copy analysis to find out whether lhs and rhs are aliased or not.
But for some code type, there is no rhs variable or the copy is always needed. Also, the post-deallocation code of these code types have the pattern. Thus, we added new 'postProcessor' to process those code types.